# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravue API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "email@email.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/avatars/{avatarId}": {
            "get": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Get an avatar",
                "operationId": "getAvatar",
                "parameters": [
                    {
                        "name": "avatarId",
                        "in": "path",
                        "description": "Avatar id to be shown",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avatar"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Update an avatar",
                "operationId": "updateAvatar",
                "parameters": [
                    {
                        "name": "avatarId",
                        "in": "path",
                        "description": "Avatar id to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update avatar image",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avatar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avatar"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Delete an avatar",
                "operationId": "deleteAvatar",
                "parameters": [
                    {
                        "name": "avatarId",
                        "in": "path",
                        "description": "Avatar id to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/avatars": {
            "post": {
                "tags": [
                    "Avatars"
                ],
                "summary": "Create an avatar",
                "operationId": "createAvatar",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avatar"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "title": "Api response",
                "description": "Api response",
                "properties": {
                    "code": {
                        "description": "Status code",
                        "type": "integer",
                        "default": 200
                    },
                    "type": {
                        "description": "Response type",
                        "type": "string"
                    },
                    "message": {
                        "description": "Reponse message",
                        "type": "string",
                        "default": "API message response"
                    }
                },
                "type": "object"
            },
            "Avatar": {
                "title": "Avatar model",
                "description": "Avatar model",
                "properties": {
                    "id": {
                        "description": "Avatar ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "string",
                        "default": 4
                    },
                    "file_name": {
                        "description": "File name",
                        "type": "string"
                    },
                    "media_type": {
                        "description": "Media type",
                        "type": "string"
                    },
                    "data": {
                        "description": "Data base64",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Example: Bearer 1|tYfXgGERi0gcGH7Vdu6TYdWdijCCdrdPjY6jGBphAnwx",
                "in": "header",
                "bearerFormat": "1|tYfXgGERi0gcGH7Vdu6TYdWdijCCdrdPjY6jGBphAnwx"
            }
        }
    }
}